consolidated the playlist menu into a separate component
